% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testcov1.R
\name{testcov1}
\alias{testcov1}
\title{One-Sample Covariance Test}
\usage{
testcov1(data, Sigma0 = diag(ncol(data)), a0 = 2, b0 = 2,
  gamma = 1)
}
\arguments{
\item{data}{an \eqn{(n\times p)} data matrix where each row is an observation.}

\item{Sigma0}{a \eqn{(p\times p)} given covariance matrix.}

\item{a0}{shape parameter for inverse-gamma prior.}

\item{b0}{scale parameter for inverse-gamma prior.}

\item{gamma}{non-negative variance scaling parameter.}
}
\value{
a named list containing: \describe{
\item{log.BF.mat}{a \eqn{(p\times p)} matrix of pairwise log Bayes factors.}
}
}
\description{
One-Sample Covariance Test
}
\examples{
\dontrun{
## generate data from multivariate normal with trivial covariance.
data = matrix(rnorm(100*5), nrow=100)

## run test with different parameters
out1 = testcov1(data)
out2 = testcov1(data, a0=5.0, b0=5.0) # change some params

## visualize two Bayes Factor matrices
par(mfrow=c(1,2), pty="s")
image(exp(out1$log.BF.mat)[,5:1], main="default")
image(exp(out2$log.BF.mat)[,5:1], main="a0=b0=5.0")
}

}
