% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testcov2.R
\name{testcov2}
\alias{testcov2}
\title{Two-Sample Covariance Test}
\usage{
testcov2(X, Y, a0 = 2, b0 = 2, gamma = 1)
}
\arguments{
\item{X}{an \eqn{(n_1\times p)} data matrix where each row is an observation.}

\item{Y}{an \eqn{(n_2\times p)} data matrix where each row is an observation.}

\item{a0}{shape parameter for inverse-gamma prior.}

\item{b0}{scale parameter for inverse-gamma prior.}

\item{gamma}{non-negative variance scaling parameter.}
}
\value{
a named list containing: \describe{
\item{log.BF.mat}{a \eqn{(p\times p)} matrix of pairwise log Bayes factors.}
}
}
\description{
Two-Sample Covariance Test
}
\examples{
\dontrun{
## generate X from standard, isotropic normal distribution
## generate Y where the last column is only differed.
X = matrix(rnorm(200*5), nrow=200)
Y = cbind(matrix(rnorm(50*4),nrow=50),rnorm(50,sd=5))

## run test with different parameters
testcov2(X, Y)
testcov2(X, Y, a0=5.0, b0=5.0) # change some params
}

}
