% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testmean2.R
\name{testmean2}
\alias{testmean2}
\title{Two-Sample Mean Test}
\usage{
testmean2(X, Y, a0 = 2, b0 = 2, gamma = 1)
}
\arguments{
\item{X}{an \eqn{(n_1\times p)} data matrix where each row is an observation.}

\item{Y}{an \eqn{(n_2\times p)} data matrix where each row is an observation.}

\item{a0}{shape parameter for inverse-gamma prior.}

\item{b0}{scale parameter for inverse-gamma prior.}

\item{gamma}{non-negative variance scaling parameter.}
}
\value{
a named list containing: \describe{
\item{log.BF.vec}{a length-\eqn{p} vector of log Bayes factors.}
}
}
\description{
Two-Sample Mean Test
}
\examples{
\dontrun{
## generate X from standard, isotropic normal distribution
## generate Y where the last column is only differed.
X = matrix(rnorm(200*5), nrow=200)
Y = cbind(matrix(rnorm(50*4),nrow=50),rnorm(50,mean=5))

## run test with different parameters
testmean2(X, Y)
testmean2(X, Y, a0=5.0, b0=5.0) # change some params
}

}
